+++ src/cmdtable.cpp
@@ -40,7 +40,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "change", "remove", },
         2,
         "Command Description: Sets the Server's Message of the Day.",
-        "Command Usage: /motd change | remove *Message. "
+        "Command Usage: /motd [ change | remove ] *Message. "
         "e.g. (/motd change No cheating!) or (/motd remove remove) to disable the MotD.",
         GMRanks::CoAdmin,
         true,
@@ -49,10 +49,10 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
     {   //Server UpTime, Connected Users, Connected Admins.
         { "info" },
         1,
-        { "server", "soul" },
-        2,
+        { "server", "soul", "muted", "quarantined" },
+        4,
         "Command Description: Shows the Server or User Information.",
-        "Command Usage: /info server | soul e.g. \"/info soul 4000\".",
+        "Command Usage: /info [ server | soul | muted | quarantined ] e.g. \"/info soul 4000\".",
         GMRanks::GMaster,
         true,
         GMCmds::Info,
@@ -63,7 +63,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "soul", "ip", "all"  },
         3,
         "Command Description: Bans the selected user and prevents their future connection to the server.",
-        "Command Usage: /ban Soul | IP | All <#>s(Seconds) | m(Minutes) | h(Hours) | d(Days) *<Reason(Optional)>. "
+        "Command Usage: /ban [ Soul | IP | All (Permission Required) ] <#>s(Seconds) | m(Minutes) | h(Hours) | d(Days) *<Reason(Optional)>. "
         "(Duration is default to 30 Days if not provided). e.g. (/ban soul 4000 Bad Soul!) or (/ban soul 4000 30m Bad Soul!)",
         GMRanks::CoAdmin,
         true,
@@ -75,7 +75,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "soul", "ip", "all" },
         3,
         "Command Description: Removes a ban from the selected user and reallows them to connect.",
-        "Command Usage: /unban Soul | IP | All (Permission Required) *Reason (Optional). "
+        "Command Usage: /unban [ Soul | IP | All (Permission Required) ]*Reason (Optional). "
         "e.g. (/unban soul 4000 Good behavior) or (/unban ip 10.0.0.1 Good behavior.)",
         GMRanks::CoAdmin,
         true,
@@ -87,7 +87,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "soul", "ip", "all" },
         3,
         "Command Description: Disconnects the selected user from the server.",
-        "Command Usage: /kick Soul | IP | All (Permission Required) *Reason (Optional). "
+        "Command Usage: /kick [ Soul | IP | All (Permission Required) ]*Reason (Optional). "
         "e.g. (/kick soul 4000 Booted!) or (/kick ip 10.0.0.1 Booted!.)",
         GMRanks::GMaster,
         true,
@@ -99,7 +99,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "soul", "ip", "all" },
         3,
         "Command Description: Adds a network mute to the selected User.",
-        "Command Usage: /mute Soul | IP | All (Permission Required) <#>s(Seconds) | m(Minutes) | h(Hours) | d(Days) *<Reason(Optional)>. "
+        "Command Usage: /mute [ Soul | IP | All (Permission Required) ] <#>s(Seconds) | m(Minutes) | h(Hours) | d(Days) *<Reason(Optional)>. "
         "(Duration is default 10 Minutes if not provided). e.g. (/mute soul 4000 Bad Soul!) or (/mute soul 4000 30m Bad Soul!)",
         GMRanks::GMaster,
         true,
@@ -111,11 +111,35 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "soul", "ip", "all" },
         3,
         "Command Description: Removes a network mute imposed on the selected User.",
-        "Command Usage: /unmute Soul | IP | All (Permission Required) *Reason (Optional). "
+        "Command Usage: /unmute [ Soul | IP | All ] (Permission Required) *Reason (Optional). "
         "e.g. (/unmute soul 4000 Good behavior) or (/unmute ip 10.0.0.1 Good behavior.)",
         GMRanks::GMaster,
         true,
         GMCmds::UnMute,
+    },
+    {   //Command Implemented.
+        { "quarantine" },
+        1,
+        { "soul", "ip" },
+        2,
+        "Command Description: Adds a network quarantine to the selected User. Quarantined Users can only interact with other Quarantined Users.",
+        "Command Usage: /quarantine [ Soul | IP ] *Reason (Optional). "
+        "e.g. (/quarantine soul 4000 Good behavior) or (/quarantine ip 10.0.0.1 Good behavior.)",
+        GMRanks::GMaster,
+        true,
+        GMCmds::Quarantine,
+    },
+    {   //Command Implemented.
+        { "unquarantine" },
+        1,
+        { "soul", "ip" },
+        2,
+        "Command Description: Removes a network quarantine to the selected User. Quarantined Users can only interact with other Quarantined Users.",
+        "Command Usage: /unquarantine [ Soul | IP ] *Reason (Optional). "
+        "e.g. (/unquarantine soul 4000 Good behavior) or (/unquarantine ip 10.0.0.1 Good behavior.)",
+        GMRanks::GMaster,
+        true,
+        GMCmds::UnQuarantine,
     },
     {   //Command Implemented.
         { "msg", "message" },
@@ -123,7 +147,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "soul", "ip", "all" },
         3,
         "Command Description: Sends a message to the selected User.",
-        "Command Usage: /msg Soul | IP | All (Permission Required) *Message. "
+        "Command Usage: /msg [ Soul | IP | All ] (Permission Required) *Message. "
         "e.g. (/msg soul 4000 Hello.) or (/msg ip 10.0.0.1 Hello.)",
         GMRanks::GMaster,
         true,
@@ -168,7 +192,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "stop" },
         1,
         "Command Description: Initiates server shutdown in <n>s(Seconds) | m(Minutes) | h(Hours) | d(Days) (30 Seconds if duration is not provided).",
-        "Command Usage: /shutdown <n>s(Seconds) | m(Minutes) | h(Hours) | d(Days) *<Reason(Optional)>. e.g. (/shutdown), (/shutdown Seconds 30) "
+        "Command Usage: /shutdown [ <n>s(Seconds) | m(Minutes) | h(Hours) | d(Days) *<Reason(Optional)> ]. e.g. (/shutdown), (/shutdown Seconds 30) "
         "will cause the server to shutdown in 30 seconds, (/shutdown stop) will cease an inprogress shutdown.",
         GMRanks::Owner,
         true,
@@ -180,7 +204,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "stop" },
         1,
         "Command Description: Initiates server restart in <n>s(Seconds) | m(Minutes) | h(Hours) | d(Days) (30 Seconds if duration is not provided).",
-        "Command Usage: /restart <n>s(Seconds) | m(Minutes) | h(Hours) | d(Days) *<Reason(Optional)>. e.g. (/restart), (/restart 30s) "
+        "Command Usage: /restart [ <n>s(Seconds) | m(Minutes) | h(Hours) | d(Days) *<Reason(Optional)> ]. e.g. (/restart), (/restart 30s) "
         "will cause the server to shutdown in 30 seconds, (/restart stop) will cease an inprogress restart.",
         GMRanks::Admin,
         true,
@@ -214,7 +238,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "hide", "show", "status" },
         3,
         "Command Description: Makes the Admin invisible to others. Poof!",
-        "Command Usage: /vanish hide | show | status. When no sub-command is entered the command acts as an on|off toggle.",
+        "Command Usage: /vanish [ hide | show | status ]. When no sub-command is entered the command acts as an on|off toggle.",
         GMRanks::GMaster,
         true,
         GMCmds::Vanish,
@@ -238,7 +262,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         "Command Description: Using \"lock\" prevents other Players from entering a scene hosted by you and \"unlock\" reverses that limitation. "
         "If enabling \"allowcurrent\" then only Players online when the Scene was created can enter and the command \"allowall\" reverses that limitation. "
         "Using the command syntax \"/camp allow soul 4000\" or \"/camp remove soul 4000\" you may exempt specific Players from any limitations.",
-        "Command Usage: /camp lock | unlock | allowcurrent | allowall | allow. Remote Administrators may append \"soul *PlayerSerNum\" to override a "
+        "Command Usage: /camp [ lock | unlock | allowcurrent | allowall | allow ]. Remote Administrators may append \"soul *PlayerSerNum\" to override a "
         "Player's status. e.g. \"/camp lock soul 4000\".",
         GMRanks::User,
         true,
@@ -250,7 +274,7 @@ const QVector<CmdTable::CmdStruct> CmdTable::cmdTable =
         { "create", "join", "kick", "leave" },
         3,
         "Command Description: ",
-        "Command Usage: /guild *message | create | join | kick | leave. e.g. (/guild *chat), (/guild create *Name) ",
+        "Command Usage: /guild [ *message | create | join | kick | leave ]. e.g. (/guild *chat), (/guild create *Name) ",
         GMRanks::User,
         false,
         GMCmds::Guild,